Been working on formform.calc where I was able to implement most of the functions that I previously had in the rescript branch and add some new ones or generalize/extend old ones.
I have been trying to unify the “constant” and “formDNA” concepts by
just using keywords, since constants can be considered formDNA of
dimension 0 and formDNA can be used just like a constant in FORMs. I am
unsure yet if it was a good idea, since performance penalty of
conversion between keywords and sequences might turn out to outweigh the
benefits over just using const seqs for formDNA. However, I still like the
compactness of keywords/strings over seqs and that they might be more
easily parsed/differentiated against FORMs. But this is representation
vs implementation and there may be a way around it or it may actually be
insignificant because of how keywords/strings are implemented?

However, my “dna-seq” format tries to circumvent this indecision for now
by providing a more general structure for transformations which is also
useful for conversion to/from sort-codes and to enable different const
representations which formDNA would usually not allow, but may be useful
in some cases.

Although the `dna->quaternary` function returns a string, the resulting
number cannot be parsed as a big integer from ClojureScript without
interop because it turns out that `read-string` just parses normal js
numbers regardless of the size (in contrast to Clojure). Have to keep
that in mind, but atm only `make-compare-dna` uses BigInts to compare formDNA.

`expand-dna-seq` is an example of a new function that could turn out to
be quite useful in research (I want to generalize it with `shrink-dna-seq`
which, contrary to creating overspecificity in expand, will remove specificity).
Another example are the new `rel` and `inv` functions which can now also
relate and invert formDNA of different dimensions.