For `formform.expr`, I have adapted/extended the evaluation and reduction functions from my `formform-clj-lab` experiment 3. The two reduction functions work together, reducing content and context (in FORMs) in accordance with the axioms of FORM logic. They can be used independently of `eval-expr` to simplify any expression, while any irreducible part remains as is. This could not be done with my previous implementations and I plan to even represent partial information in formDNA with something like “holes”. formDNA should also be reducable in expressions, but I still have to find out if it can be done with this approach. All `eval-expr` does is to determine the value/dna that the reduced FORM indicates. I would like to make this compose, such that evaluations can be part of other expressions, thus my previously stated intentions with formDNA.
Added an option map to `permute-dna-seq` and `dna-seq-perspectives` to
switch off the dimension limit since it might sometimes be useful when
computational resources allow for processing such enormous collections.
One can also push the limit by writing the output in a file instead of
directly to the REPL.

I also wrote vdict generators (which may be useful for value tables) and
a `vdict->vmap` function that creates vmaps as nested maps. I am unsure
about this format, since it seems wasteful considering they could just
be vectors and have numeric indices determine each value, but it might
turn out to be useful. Turns out sorting maps is really slow though, so
while this makes sense for vdicts (optionally), not so much for vmaps.