I started experimenting with Clojure specs since I need more reliable tests on the periphery of my functions. So I “spec-ified” some of my predicates and used them in function specs to validate input and output data. This allows me to `instrument` them and all my previously written unit tests will automatically get property-based testing as well.
I can also use specs for generative testing using `stest/check`, but often
this requires writing generators manually which seems to be tricky in
some cases and I feel like for most of my functions I get much more out
of unit tests on specific examples that I can manually verify. But it
might be valuable to have generators on important data structures to be
able to generate random data and explore their generative space to catch
some edge cases.

Specs and generators can be easily integrated in normal functions, so I
can reuse them in `x?` and `rand-x` functions (it is unclear if there is
a performance penalty though). For production builds, I can use
`s/valid?` and `s/conform` for input validation and something like
`s/explain` to get useful error messages (although I might have to
somehow transform them to be meaningful to the user). However, for now I
will only use input validation where it is really necessary or in user
interfaces to not drag down the performance too much. I might add some
asserts here and there, but they can be ignored in production.

Besides specs, I have experimented with representation of re-entry
FORMs. I added two macros for nested sequences from my experiments repo
and wrote 6 other macros that generate all the self-equivalent re-entry
patterns as described in uFORM iFORM using my re-entry signatures in
the var names. They will be represented as normal nested expressions
with a `:f*` at the re-entry point that refers to the expression itself.

All macros build maps that represent equations where the keywords are
symbols equated to their expressions which might be reused in expressions
in case of recursion. This structure is actually equivalent to how envs
are represented in evaluation and I plan to use it as a “memory FORM”
that allows me to expand the env during evaluation such that the FORM(s)
specified there are reduced in place as in an expression but also kept
in “memory”/env (after reduction) to be reused from this point on.

I am unsure yet if this might help me build a simpler algorithm for
evaluation of self-equivalent re-FORMs, but it may come in handy if one
wishes to re-enter the same FORM several times in an expression without
having to use programming to do so and without having to evaluate the
same expression multiple times. Like a user-defined symbol that locally
adds to the env. Will have to give this some more thought.