In the last weeks I have started to refine and clarify my modular DSLs and focussed on how one might use the library in practice. To express and sharpen my ideas for each module, I set up a computational notebook engine called Clerk which integrates well with the REPL and allows me to create some kind of documentation on the fly while being a space for experiments and development. In the future this can also help others to understand how they can work with the library.
I also started to provide constructors for all the data objects I
introduce, because even if their internal structure mostly consists of
simple Clojure data types like vectors and keywords, I want to go back
to hiding such implementation details so that we can be more specific and
secure when working with formform data. This is not to say that a user
cannot just work with the naked data for traversal, transforms, etc. but
they will do so knowing that it might be unsafe.

For this reason I experimented with adding a metadata-tag to expressions
that acts as a lightweight type hint which can be checked against in
functions that expect expressions and provides an additional safety layer
on top of specs. Anyone can add these manually, so this is no forced
constraint. This will not work on e.g. keywords and other primitives so
maybe I’ll have to reconsider it later on.