I further clarified and simplified the expression DSL by separating primitives, which also include FORMs and unclear FORMs (all primitives vars are in capital letters now) from expressions, which are the actual building blocks a user would work with. This hopefully clears up some confusion about how to build FORMs.
There are a number of pre-defined Expressions to wrap primitives and special
expressions for variables, constants and formDNA, which eventually reduce to
basic expressions including only primitives (not implemented for formDNA yet).

Expressions can be composed, resulting in expressions which context is merged
with the context of their inner expressions. Expressions with local envs will
not be merged, of course, but sometimes can be reduced by applying their env to
their contents, in which case the env will disappear and the context can be
merged (to be implemented).

Expressions can be abstractly defined as Clojure functions that return
expressions, following the convention to put a mid-dot `·` before the name of
the function. Other expressions can call them either directly or indirectly by
quotation (in which case the reducer will take care of calling them) (the
latter is not implemented yet).

Besides that, formDNA expressions now return a `dna` (single keyword) instead
of a `dna-seq` which means that `=>` will also do that. I might roll this back
if it turns out to be inefficient for further processing, but it reads more
nicely in the output and aligns better with expectations, since this is
currently the standard representation for formDNA (`dna-seq` is more like an
intermediate format).