Rewrote my nesting macros to use them as ordinary functions, so I can specify nested expressions in a more flexible way (I am learning that macros are more about syntactical shortcuts to be used/expanded while coding and it is bad practice to use them for transformation of non-code data). I also added a chaining transformation, just because it could be convenient for users.
I also implemented all self-equivalent re-entry FORMs in their shorthand
representation using either an options map or my re-entry signatures for
specification. With the nesting function, I was also able to write the reduce
transformation for these FORMs down to their lower-level representation as
recursive equations (implemented as a local env on the expression with the
starting variable as its only content).

My notebook for the `expr` module has also been updated and restructured with
some explanations. It has been really helpful to keep an overview and a more
practical perspective on things.