I decided to clearly separate “content” from “context” (provided by expressions) in the `expr` module. This way, there will be no more confusion about the semantics of and “expression” vs a “FORM”, because an expression really is just an arrangement of FORMs intended for evaluation. It is a context on its own and never a content, hence we only need a single expression constructor that can be used to wrap contents rather than create new content types.
I may have to create aliases for the mid-dot var names if I want formform to be
used from JS, but that would be no problem.

This setup is remarkably similar to what I have previously done in my rescript
branch of formform, except that the expression type there could not be extended
as easily. However, arriving at this “final form” in both branches indicates to
me that it might be the most clear, flexible and simple approach that I could
find to serve the purpose of the library. Alas, there are many things that I
would not have thought about were it not for the change to Clojure. It allows
me to be a lot more experimental without having to stick to a fixed structure
which is important to me at this stage in development, because I have to figure
this stuff out as I go. Maybe in the future I port this back to a strongly
typed language if I have more critical requirements, but for now Clojure serves
me well.