Added more expansion functions to expand special FORMs to expressions of simple FORMs. I started calling them “expand” instead of “reduce” now to clearly separate them from functions that reduce by application of axioms/theorems. It also indicates that the resulting expressions will most often be larger than what goes in.
They take content as input and return a context/expression, because often the
expansion may contain multiple items instead of just one. Taking in content
instead of context makes it easier to map over contexts. This can be a source
of error, which is why input checks are important, so I tried attaching tag
metadata on content too. Not sure if this is the way to go, since it will also
make it cumbersome to generate nested FORMs with ordinary Clojure functions,
(I am also thinking about parsing notation and EDN data), but it might not be
that big of a deal and adds additional safety checks, so we’ll see.

I also created expression constructors for “isolator” and “selector” FORMs
(as I call them), which have been especially important in my research.
Isolators are central to formDNA expansion and have been used by Ralf Peyn in
his publication “eigenFORM”. Selectors are my generalization of binary
selector/decision FORMs (which play a central role in uFORM iFORM) to making
undetermined selections (I still try to figure out what that means).

Moreover, some more refactoring, reorganization and annotation happened.