I realized that instead of using metadata tags to validate content, I can just use the keyword tags I have already attached to most of the special FORMs. So I now consistently tagged all of them and wrote specs based on that.
I also wrote `expand-expr` which is the only expansion from context to content.
If there is an env in the expression/context, it expands to a conjunction of
equivalences corresponding to each key-value pair in the env-map together with
the actual contents of the expression. If there is no env, it just expands to
a double-marked FORM. This function isn’t really useful at this point, but I
just wanted to pin down the semantics based on my conception of expressions.

Furthermore, I continued writing my usage notes to see how I could actually
explain this stuff to other people, because often I get so lost in the realm of
my own ideas that I lose connection with the user perspective. It might turn
out to be a tutorial at some point, but for now it just helps me personally.