Lots of changes accumulated over the last few days, this may be the longest commit message ever…
Expression-local envs are now reified as “memory FORMs” (how I call them).
These are special FORMs (like unclear FORMs or formDNA) and I have written
extensively in my Roam notebook about their semantics, which can be explored
with the expander I wrote, but is not really useful without actually performing
the “memorization” process. To represent this process in FORMs would most
likely require re-entry FORMs, but there is a co-dependency here, since the
representation of re-entry FORMs was my prime motivation behind memory FORMs.
To be able to represent `f = (f)` in FORMs, without the help of an algebraic
meta-language, and to have it unfold in time as an emulation of a cognitive
process (which will be useful in some modules I plan for formform), requires
some concept of memory. But this is a property of the evaluator and hence
cannot be represented adequately in the memory FORM itself, which only acts as
an injunction such as “remember this!”. I am still researching about what this
is supposed to mean, but memory FORMs are valid FORMs in the sense that they
can be reduced and evaluated in a clear way that does not contradict any law
of the FORM calculus and they provide a simple representation for re-entry
FORMs and maybe even lambda expressions (kind-of crude though).

Without local envs, expressions are very simple now to the point that they are
merely container/context for FORMs. This is good, because local envs had a very
specific use-case which never really justified the overhead.

The reduction procedure itself has been improved in various ways, but most
importantly it now remembers observed values from previous contexts and uses
this information to remove or add redundant content justified by (de)generation
theorems in FORM logic. This means that an expression such as `(a (b a))` can
now be reduced to `(a (b))` which was not possible before. Unfortunately it
greatly expands the number of (possibly redundant) iterations and calls, since
the observed FORMs in a context might change after their reduction, which
causes the reducer to take another pass to try to further simplify the reduced
FORMs via its updated observations and this causes the contexts of those FORMs
to propagate the observations and so a small drop can make huge waves. I might
add an option to disable (de)generation inference if the performance decrease
is noticable, especially if the goal is calculation, not algebraic reduction.

I also started using flow-storm as a debugger, which is a great help to make
sense of those complex nested reductions.

Besides that, formDNA can now be simplified in reduction, if any of its
specified variables can be determined by a constant. There are still more steps
required to fully integrate formDNA into evaluation, but this is also very
useful for filtering formDNA in multiple ways.

Memory FORM can already be reduced, but need more simplification. Of course,
there is the potential for infinite reduction which I already experienced and
I am currently thinking about a way to catch the stack overflows from recursive
definitions, so that I can generate better error messages for users and prevent
crashes. This is however a very niece problem since most users would never
explicitly define self-contradicting re-entry FORMs or do so at their own risk.

I also started reduction for self-equivalent re-entry FORMs, which simply
requires reducing all contexts in a chain and then pattern-matching on the
reduced chain to match all possible species (if not already obsolete).