I finished the reducer for self-equivalent re-entry FORMs and wrote some basic test-cases for it. There are still a lot more tests required and some more manual verification, but it works and for the first time in formform, re-entry FORMs can be reduced algebraically without having to be fully interpreted. This also enables us to transform them to binary FORMs if possible, which it is in a lot of cases (even more so with the alternative interpretation).
For clarity, I chose to use pattern-matching with Clojures core.match library,
which avoids a lot of branching through the individual cases and sub-cases and
makes it very clear even to outsiders which patterns are being identified and
what the result looks like for each one. I have also identified 6 patterns that
exhaust all cases where the alternative interpretation applies, given the nested
contexts are sufficiently reduced (needs more verification though).

For nestings and other functions that apply multiple contexts, I decided to
require all contexts to be explicitly written as vectors. Although it is more
verbose, it avoids confusion about content vs context and how to group content
but also makes it simpler to further process this data. I am unsure yet if I
should require vectors rather than any sequential, given that Clojure seems to
treat the vector/list distinction merely as an implementation detail and I
guess it would be wise to hide this fact in interfaces so users don’t have to
think about it. I’ll have to check if this distinction is leaking through in
my interfaces for expressions/contexts (tags and merging are important here).