Did a lot of cleanup and bug fixing on the expression module. I also wrote some more unit tests, which also replaced the ad-hoc tests/experiments at the bottom of `expr.clj`. I really should write some specs for this module now, since I encountered some unexpected results due to lack of specs/typechecks. A bad one was my misuse of `seq` to coerce vectors to sequences, which returns `nil` for `[]` and not `()` (which I knew, but didn’t really think about at the time). I replaced that with `sequence` instead.
All special FORM reducer now have their own function which always takes the
FORM and the environment used in reduction. Maybe I can create a modular system
for reduction and expansion of special FORMs, where you can mix and match
reducer/expander a la carte (say you don’t want to reduce unclear FORMs or
expand seq re-entry FORMs or only want to reduce pure binary FORMs), but I’ll
have to consider if that makes sense and how I can implement the selection.

I also extended `find-vars` to be able to find only specific variables and used
that for a `filter-rems` function which filters out all variables in the rems
part of memory FORMs that are not actually referenced in its context part.
Because rem-bindings can reference previous rem-bindings or shadow them, it is
necessary to find all dependencies in the (contexts of) rem values too.
This may drag down performance in some cases, but since substitutions are only
performed once per reduction and to keep re-entry definitions, this is a
necessary evil (maybe it can be optional?).