I had to make some major changes to the reduction functions to better handle edge cases like envs where multiple successive substitutions are possible or making sure that substitutions from env happen before degeneration of observed values. I believe there are still more edge cases I haven’t found yet, but more and more tests will reveal them in time, so it is an ongoing process.
Reduction of memory FORMs also had some hairy cases related to that, since the
env and observed data must be integrated with rem value reduction, shadowing
and so on. It is important to carefully consider what the logically correct way
of reduction needs to be in each case, especially if multiple interdependencies
between references, nested memory FORMs and self-reference are involved. I don’t
think I have all the answers yet, but it is much more consistent and robust now.

Similar to seq-reentry FORMs I also made sure that rem values in memory FORMs
are always wrapped in a context/sequential, so there is no ambiguity later on,
especially when ordinary FORMs and contexts become indistinguishable. The
interface still allows (similar to envs) to provide values as content and in
fact one must explicitly construct expressions for contexts (using `··`). I am
unsure about this distinction yet, but it feels more natural and simpler to
write than having to wrap every value in a vector (which can easily cause bad
mistakes when you forget that). It is also more aligned with how it works in
env maps and more ergonomic, since most of the values we use here are simple
constants and only rarely multiple contents.