Implemented expression parsing for formula notation using the Clojure library “instaparse”, which makes writing parsers almost too easy. I put the (extended) Backus-Naur form in a separate file for convenience and used the transform function from instaparse to specify how to transform the syntax tree into a formform expression. Given the syntax tree has a clean structure this is almost trivial to do.
However, there are still many edge-cases to test for and my main problem
right now is how to deal with separation between values and variables,
for which I may need some kind of lookahead in the ebnf. For example,
the formulae `a:M`, `:Ma`, `1a` and `a1` should not be valid since they
are ambiguous.

I also want to write the inverse function to transform expressions into
strings in formula notation. But in order to do that, I will need to
find a clear extension for memory FORMs.