Solved a bunch of problems with the parser, especially the separation between literals (variables/values not wrapped in “something” - not sure about this term yet), which should not be optional as between literals and nodes (“wrapped things”). My solution is a bit crude, but it works.
Also separated the formula parser from the expression module, which
makes writing tests much more pleasent (my neovim is slow on 1k+ loc
files, maybe due to some plugins I use). It also makes sense to separate
them logically, because formula notation is not strictly necessary to
write expressions in formform.

There is still the issue with unwanted optional whitespace between
components of constants and formDNA, which is due to the (experimental?)
auto-whitespace feature in instaparse. It works well if everything is a
token, but in these cases where I decided to separate the prefix from
the actual value, so the prefix will have unwanted optional whitespace.
Maybe I should insert optional whitespace manually in the ebnf instead
of relying on the auto-whitespace feature from instaparse.