Fixed the unwanted whitespace between components of constants and formDNA via strict tokenization. A formDNA can now be written without brackets if it does not have a varlist attached, but must be wrapped in brackets if it does. I think this is a good compromise since constants like `:U` are just 0-dimensional formDNA without a varlist attached (which is redundant for it will be empty), so something like `:NUIM` will just be interpreted as having a generic variable name for its dimension that doesn’t conflict with the variable names in most expressions (such as `v__0`). And since those will be consistent between generic formDNA, they can be related algebraically if one wishes to examine formDNA relations more generally. Otherwise, a term like `[[x] :NUIM]` can be written to name variables explicitly (spaces after the varlist are allowed now).
The grammar as well as the transformations are pretty complete now and I added
a lot more tests to make sure everything works as expected. I haven’t decided
yet about what to do with memory FORMs, but they will get some kind of notation
that is consistent with the rest and easy to write.

For the calc module, I have been dissatisfied with my dna <-> dna-seq
converters, so I decided to abstract the converter functionality so that one
can easily create additional converters for arbitrary sequence types that can
always reorder between sort codes. This makes the existing code much dry-er.