Created a function that attaches geometric information to the vmap structure as it is built (via passing a transformation function to `vdict->vmap`). I heavily refactored my old vmap constructor from formform 1 for this purpose and simplified it quite a bit, though I still have to clean up some parts and rethink what options I want.
It builds a tree of maps that is self-similar on all levels so that
subtrees could be detached or combined and still be geometrically
correct. This makes it flexible enough to work with different
visualization formats, such as SVG, which I have already tested in a
Clerk notebook.

I am not sure whether I want to keep it in the formform core or create
an extension library that also includes geometric constructors for
expressions (hook-style, circular or whatever). Maybe I start by moving
it to a different namespace, since I don’t want to branch out too much
at this point.