I started to rename the source files to .cljc for a common interface between Clojure and ClojureScript with formform. Testing this from ClojureScript revealed some incompatibilities, most of which were due to native Java code that could be easily fixed with reader conditionals. However, I did not anticipate that the mid-dot character `·` is not allowed in ClojureScript and therefore had to find a different syntactic device to mark formform expressions. My choice was the angular brackets, which still provide a concise and somewhat iconic naming scheme.
A FORM expression is just `<>` and every other expression adds to that,
like `<uncl>` for unclear FORMs. I decided to make use of my special
FORM notation (still wip) to designate variations of `<>` that act as
notational shortcuts for simple patterns:
- `<->` expands to `(<> (<> … …))`
- `<|>` expands to `(<> …) (<> …)`
- `<+>` expands to `(<> (<> …) (<> …))`

Nesting operators build on these notational semantics but their brackets
point in a single direction, like `<-<` or `>>`. The dash in between can
be thought of as negation or subtraction but is really a connection that
binds the contents of the FORM to another (sub-)FORM. The vertical line
can be thought of as separation/division but is really a conjunction of
the contents of the FORM. The plus can be thought of as a combination of
those two operations.

I am unsure if this will be clear and intuitive to beginners, I may have
to provide a good explanation or fall back to a different naming scheme
altogether. However, most people will want to use formula notation anyway
without any of these specialties, so they won’t be bothered by that.

However, I’ll really miss the simplicity and inconspicuousness of the `·`
approach, it was much easier to read and write. But in the end this is
only programming syntax and doesn’t really matter that much and I may
even change to words in the future since angular brackets are not
allowed on the pure JavaScript side. But maybe there is a simpler
alternative to build expressions – I’ll have to rethink what I’m doing.