I was dissatisfied with my previous approach to create “special FORMs”, so I started to re-imagine the whole API with a more unified and flexible approach that follows my FORM-as-data principle more consistently.
Now, there are only 4 primitives in the `expr` module:
- `(…)` or `[…]` for simple FORMs
- `"name"` or `'name` for variables
- `:sym-name` for predefined symbols
- `[:op-name …]` for predefined operators

Every “special FORM” is a predefined operator. Symbols and operators can be
defined using the macros `defoperator` and `defsymbol` by specifying a keyword
as a unique id which the macros will “register” via generic dispatch in
multimethods. This enables functions to know about these FORMs without having
to add their interpreters, simplifiers or constructors manually. I decided
against using records and protocols here since I find multimethods to be a more
flexible, functional and interop-friendly approach to polymorphism, which
doesn’t rely on host-specific OOP constructs.

I might get rid of `op-get` though, since the `defoperator` macro will define
methods for it for all parameters of the expression, which may bloat memory
unnecessarily while `op-data` can do the same (albeit a little less efficient).

The bare minimum to define an operator is for the user to define its
interpretation through `interpret-op`. E.g. `:and` can be defined with an
interpretation of `[[x] [y]]` and used like `[:and "red" "green"]`.
Operators can have any shape, but must start with a keyword, so for example
`[:mark-n n x]` might be an operator that generates a FORM with `x` marked `n`
times or `[:sel3-inv [x [y] z]]` might invert a selection to `[[x] y [z]]`.
Variable numbers of parameters are also possible, since we are just working
with normal Clojure functions.

This flexibility enables the user to be very expressive with FORMs and define
their own “macros” as shortcuts for complex constructions, which can also
be composed and combined at will. They can also define a custom simplification
procedure via `simplify-op` that will then be preferred by the simplifier over
strict interpretation. It can manipulate the environment during simplification
and do all sorts of unsafe and magical stuff, so the user is responsible to
verify the validity and correctness of their approach or just use it for crazy
experiments. Self-equivalent re-entry FORMs, formDNA FORMs and memory FORMs use
this feature to implement their special behavior.

The former “expression” primitive `[…]` has also been replaced by an operator,
since every FORM or content is considered an expression now (no arbitrary
distinctions anymore!). `[:- …]` is just called an arrangement and interpreted
as `[[…]]`. It is not required to build an expression and can be used for
grouping and unmarked relations. I also introduced `:*` and `:|` as syntactic
sugar for `(a)(b) …` and `((a)(b) …)`, which are ideas from my thoughts on
special FORM syntax. They may or may not be useful, we’ll see.

Construction of operators can be directly as vectors (no metadata needed) or
more safely through `(make :op args…)`, provided the operator is registered
for `make-op`. `(form …)` is a corresponding constructor for simple FORMs and
`(seq-re :sign nested-exprs…)` constructs self-equivalent re-entry FORMs. In
the REPL, I would use “just data” while in more critical functions/definitions
constructors would be the safer choice, since they can (soon) validate inputs.

I also decided to get rid of `:mn`, since it is just another way to write `:U`
and it is annoying to have to recognize and test both of them in various
functions. If a user wants `:mn`, they can define it as an expression symbol
that gets interpreted as `:U`.

This might be the longest commit ever and I most definitely missed some changes,
but I felt that I needed to document the rewrite of `expr` from a conceptual
side before moving on. The old draft is still around as I still need to adapt
and add many things (and tests), but it will be deleted later.