Last month I decided against using single keywords for formDNA and instead for using a sequence of constants as the only valid representation. Although it is less compact and differs from my notational representation of formDNA, there is a strong point against very large keywords, which is the fact that Clojure seems to cache keywords and it may bloat the application state if large quantities of high-dimensional formDNA would be passed around. The Clojurian slack community made me aware of this issue and so I decided to rather take advantage of having only the four constants as cached keywords.
I have now applied this change to the `calc` module and it greatly simplified
the interface since I mostly had to convert to/from keywords which actually was
a bad decision in hindsight. There is no explicit `dna-seq` definition anymore
since this more general structure is not needed anymore, but I keep on using
the term for some functions which operate independent of whether the elements
of the sequence are constants or e.g. numbers.

Also fixed a lot of things in `expr` and changed tests for formDNA to reflect
my changes. Additionally, I created two Clerk notebooks to see if everything
makes sense conceptually. Still a lot of things have to be cleaned up and
fixed until I can release an alpha version, but the `expr` module is now in a
state that is almost as correct as it was before the big changes.