Finally removed the old expression module backups since they are not needed anymore.
Modified the parser grammar to allow all symbols to be used in formula notation
via colon prefix. Constants are also just processed as expression symbols, as
they are in the new `expr` module. This also allows to use digits instead, like
`:3`, which will be interpreted as `:M` (or `:I`, when input is in NMUI code).

Since formDNA and constants have different representations in `calc`, I also
reflected this change in the grammar, such that `:U` will be parsed as a
constant, `:UUUU` as a symbol (unregistered, maybe) and `::UUUU` as a formDNA
expression with an arbitrary variable/term. Yes, the double-colon is back,
but this differentiation still makes sense and is already established. Since
no one really uses formDNA besides myself for now, I may not change the prefix
for NUIM code interpretation, but I am not sure yet.

For operators, I would like them to also be usable from formula notation.
This will require some adjustments, but I think `[:op …]` will be fine here too
and would also work for formDNA, etc.