Extended and refactored the formula grammar to be able to parse operators. At the moment, non-predefined operators will only work with expression content, so they must follow the parsing rules of expressions. I might add numbers to the top level for things like iterative interpretation, but I don’t really want to add support for any Clojure form since this could add a lot of unwanted side-effects and make the parsing more brittle. If people want more flexibility, formulas aren’t really the right tool and they should work with data instead.
Predefined operators are read with the parsing rules of their contents, which
are the same as for their syntactic shortcuts, so they don’t apply to the
operators themself but to the respective “special content”, like varlists,
re-entry signatures or formDNA. So I might add such content to non-predefined
operator parsing as well, but users wouldn’t be able to add more special content.

I also added memory FORM parsing to operators with remlists as special content
which parses as `<x1 = y1>, <x2 = y2>, … | <exprs>` instead of the tuple-vector
representation. I am thinking about using this exact notation outside of the
operator before any expression to be able to quickly define associations, so
that it becomes another syntactic shortcut, but not sure if it is a good idea.

Out of curiosity, I also added operators to `expr` that define what I call
`logical spaces`, but I might not keep them, especially since this is ongoing
research.