Fixed some ClojureScript-specific bugs and added some reader-conditionals to make macros and external libs work with cljs. Since instaparse does not work with separate files for grammar definitions in cljs and I do not want to hand-escape quotes and slashes for every grammar change, I made a Babashka script to do the work for me, which will create a .cljc file with the same name as the grammar file and puts the definition in a `defparser` call there. I went a little overboard with this one, but now it can be used more generally and not just on formula.ebnf.
I also changed some of the function names in `expr` and `io` to default to more
descriptive names for the API and make aliases to save some typing. This may
also help if functions need to be called from JavaScript, since many symbol
characters are not allowed there.

Not sure if the previous `eval-expr` and `eval-all` in the expression module
are really needed, so I commented them out for the time being. Maybe I’ll find
some good use-cases which justify the differentiation later on.