Improved specs for the `calc` module and added function specs for all relevant functions there. Some of them may be too fine-grained, some may need more spec-ification, but it works for now.
I also improved and added some of the generators, but not sure if it is worth
the effort to create custom generators for vdicts, vmaps, etc. since I do not
believe that generative tests provide much value for formform. But it may be
nice to get functions that can generate random samples of these types.

Instrumentation of the function specs for now will only be used in tests, where
they can provide the value of asserts and even guarantees similar to a static
type system, but with much more specificity and flexibility. I’ll try to use
the `orchestra` library to also check return values next time.

In the process I also imrpoved the `make-dna` function, which is much more
useful and flexible now and lets you combine multiple nested dna-seqs into one
flat formDNA. Will have to see how useful this is though.

I also got rid of the vars for `:N :M :U :I` in `calc`, since they are not
really useful abstractions. If I’d ever change the representation of constants
(which is very unlikely), I would have to do a lot more work than just replacing
occurrences. There are also many places (e.g. in `case` forms) where I can only
use the keyword literals, which creates a lot of inconsistency. Not worth the
overhead.