Swapped `clojure.spec.test` with the `orchestra` library, which allows me to fully instrument function specs in tests, so return values are also checked. It is of course slower, but already uncovered some problems in my spec definitions that I corrected. I especially fixed the `make-compare-dna` function that was broken with the switch to sequence representation of formDNA. It cannot compare collections of formDNA anymore, but I am pretty sure that this is not that useful anyway.
For the `expr` module, I introduced a new `::valid-operator` spec that matches
operators like `::operator`, but also checks the op’s arguments if it has a
method registered for `op-spec`, which returns a spec for that operator.
It works via spec’s `multi-spec` function and allows me to validate registered
operators in more detail while not having to change the spec when I add/remove
operators. Maybe it even allows users of the library to add their own specs
for their custom operators. Conforming to those specs can also extract useful
information about the argument names.

I also exchanged keyword literals of the most common operator symbols with vars
to be able to more easily change them if needed.